Develop

+++ company/src/main/java/ru/scarlet/company/controllers/FileController.java
@@ -10,8 +10,13 @@ import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import ru.scarlet.company.client.AuthClient;
 import ru.scarlet.company.client.FileServiceClient;
 import ru.scarlet.company.dtos.ErrorDetails;
+import ru.scarlet.company.dtos.UsernameFromToken;
+import ru.scarlet.company.entities.FileData;
+import ru.scarlet.company.excpetions.BadRequest.BadRequestExceprion;
+import ru.scarlet.company.excpetions.Null.HeaderNullException;
 import ru.scarlet.company.services.FileService;
 
 import java.time.Instant;
@@ -24,21 +29,43 @@ import java.util.UUID;
 public class FileController {
     @Autowired
     private FileServiceClient fileServiceClient;
+    @Autowired
+    private AuthClient authClient;
 
     @Autowired
     private FileService fileService;
 
     @PostMapping("/")
-    ResponseEntity<?> addFile(@RequestParam("file") MultipartFile multipartFile){
+    ResponseEntity<?> addFile(@RequestParam("file") MultipartFile multipartFile, @RequestHeader String authToken, HttpServletRequest request){
+        if (authToken == null || authToken.isBlank() || authToken.isEmpty()){
+            throw new HeaderNullException("Header authToken is null");
+        }
         log.info("addFile ");
-        var file = fileServiceClient.addFile(multipartFile);
+
         log.info("file addded");
-        fileService.save(file.getBody());
-        return file;
+        ResponseEntity<UsernameFromToken> username = authClient.getUsername(authToken);
+        if (username.getStatusCode().is2xxSuccessful()) {
+            var file = fileServiceClient.addFile(multipartFile);
+            fileService.save(file.getBody(), username.getBody().getUsername());
+            return file;
+        } else return ResponseEntity.badRequest().body(
+                new ErrorDetails(Instant.now().toEpochMilli(),
+                        request.getRequestURI(), "Not found",
+                        400, MDC.get("CorrId")));
     }
 
     @GetMapping("/{oguid}")
-    ResponseEntity<?> getFile(@PathVariable UUID oguid, HttpServletRequest request){
+    ResponseEntity<?> getFile(@PathVariable UUID oguid, HttpServletRequest request, @RequestHeader String authToken){
+        if (authToken == null || authToken.isBlank() || authToken.isEmpty()){
+            throw new HeaderNullException("Header authToken is null");
+        }
+        String username = getUsernameFromToken(authToken);
+        FileData fileFromDB = fileService.getById(oguid);
+        if (!fileFromDB.getCreatedBy().equals(username))
+            return ResponseEntity.badRequest().body(
+                    new ErrorDetails(Instant.now().toEpochMilli(),
+                            request.getRequestURI(), "FORBIDDEN",
+                            403, MDC.get("CorrId")));
         ResponseEntity<Resource> file = fileServiceClient.getFile(oguid);
         if (file.getStatusCode().is2xxSuccessful())
             return file;
@@ -49,14 +76,27 @@ public class FileController {
     }
 
     @DeleteMapping("/{oguid}")
-    ResponseEntity<?> deleteFile(@PathVariable UUID oguid, HttpServletRequest request){
+    ResponseEntity<?> deleteFile(@PathVariable UUID oguid, HttpServletRequest request, @RequestHeader String authToken){
+        if (authToken == null || authToken.isBlank() || authToken.isEmpty()){
+            throw new HeaderNullException("Header authToken is null");
+        }
         ResponseEntity<Void> voidResponseEntity = fileServiceClient.deleteFile(oguid);
         if (voidResponseEntity.getStatusCode() == HttpStatusCode.valueOf(204)){
+            String deletedBy = getUsernameFromToken(authToken);
+            fileService.delete(oguid, deletedBy);
             return voidResponseEntity;
         }
         else return ResponseEntity.badRequest().body(
                 new ErrorDetails(Instant.now().toEpochMilli(),
                         request.getRequestURI(), "Not found",
-                        400, MDC.get("CorrId")));    }
+                        400, MDC.get("CorrId")));
+    }
 
+    private String getUsernameFromToken(String token){
+        ResponseEntity<UsernameFromToken> response = authClient.getUsername(token);
+        if (response.getStatusCode().is2xxSuccessful()){
+            return response.getBody().getUsername();
+        }
+        else throw new BadRequestExceprion();
+    }
 }
